Produces a latex longtable

Produces a latex longtable

lTab(x, label = NULL, caption.firsthead = NULL, caption.head = NULL,
  caption.foot = NULL, caption.lastfoot = NULL,
  booktabs = .op("kLat.lTab.booktabs", "kLat.booktabs", FALSE),
  toprule = .book("kLat.toprule", booktabs, "\\toprule", "\\hline"),
  bottomrule = .book("kLat.bottomrule", booktabs, "\\bottomrule",
  "\\hline"), midrule = .book("kLat.midrule", booktabs, "\\midrule",
  "\\hline"), align = .op("kLat.lTab.align", "kLat.align", "center"),
  envir = getOption("kLat.lTab.envir", "longtable"),
  colsep = .op("kLat.lTab.colsep", "kLat.colsep", ""), coldef = .coldef(x,
  colsep), rowsep = .op("kLat.lTab.rowsep", "kLat.rowsep", ""),
  rows = .op("kLat.lTab.rows", "kLat.rows", FALSE), head = .header(x, rows),
  firsthead = NULL, foot = bottomrule, lastfoot = NULL)

Arguments

x
a data.frame or matrix to form the base of the table

label
set the table's label, defaults to an empty string

caption.firsthead,caption.head,caption.foot,caption.lastfoot
paces the caption in the firsthead, head, foot, or lastfoot respectively. It is important not to set a caption in an otherwise NULL section (although an empty string is acceptable) or strange bugs can occur. It is accaptable if the section was set by default as in head and foot. Consult the longtable documentation for a more detailed explanation of these options.

booktabs
logical value, if not set will use value of kLat.(xTab|sTab|lTab).booktabs, if not set will use value of kLat.booktabs, if not set defaults to FALSE. When TRUE toprule defaults to '\toprule', midrule to '\midrule', and botrule to '\bottomrule', when FALSE those values all default to '\hline'. Has no effect when toprule, midrule, and botrule are individually set.

toprule
sets the value for the top rule, if not set will be determined by the value of booktabs

bottomrule
sets the value for the bottom rule, if not set will be determined by the value of booktabs

midrule
sets the value for the mid rule, if not set will be determined by the value of booktabs

align
set the alignment of the environment, if not set will use value of kLat.(xTab|sTab|lTab).align, if not will use value of kLat.align, if not set defaults to 'center'

envir
set the environment for the table, if not set will use the value of kLat.(xTab|sTab|lTab).envir, if not set defaults to 'tabular', 'supertabular', and 'longtable' for xTab, sTab, and lTab respectively

colsep
separator to be used between columns (i.e. '|'), if not set will use the value of kLat.(xTab|sTab|lTab).colsep, if not set will use the value of kLat.colsep, if not set defaults to an empty string. If coldef is set this value is ignored and the separators must be specificed in the coldef

coldef
sets column definition i.e. \begin{tabular}{'align'}, if not set defaults to numeric = right, character = left

rowsep
the separaotr to be used between rows (i.e. '\hline'), if not set will use the value of kLat.(xTab|sTab|lTab).rowsep, if not set will use the value of kLat.rowsep, if not set defaults to an empty string

rows
logical value to determine if rownames are included in table, if not set will use the value of kLat.(xTab|sTab|lTab).rows, if not set will use the value of kLat.rows, if not set defaults to FALSE, if TRUE the column name for the rownames column defaults to an empty string

head
sets the value for the table header, defaults to the column names; if you set this be sure to end with '\\\\'

firsthead
header on first page of table only; defaults to header; if you set this, you are responsible for setting any \hline, \toprule, or \midrule lines

foot
sets value of the table footer, defaults to the value of botrule

lastfoot
footer on last page of table only

caption
the caption for the table, unlinke xTab and sTab, there is no caption.top or caption.bottom option in longtable

Examples

lTab(mtcars)

#> \begin{center}
#> \begin{longtable}{rrrrrrrrrrr}
#> \hline
#> mpg & cyl & disp & hp & drat & wt & qsec & vs & am & gear & carb \\
#> \hline
#> \endhead
#> \hline
#> \endfoot
#> 21 & 6 & 160 & 110 & 3.9 & 2.62 & 16.46 & 0 & 1 & 4 & 4 \\
#> 21 & 6 & 160 & 110 & 3.9 & 2.875 & 17.02 & 0 & 1 & 4 & 4 \\
#> 22.8 & 4 & 108 & 93 & 3.85 & 2.32 & 18.61 & 1 & 1 & 4 & 1 \\
#> 21.4 & 6 & 258 & 110 & 3.08 & 3.215 & 19.44 & 1 & 0 & 3 & 1 \\
#> 18.7 & 8 & 360 & 175 & 3.15 & 3.44 & 17.02 & 0 & 0 & 3 & 2 \\
#> 18.1 & 6 & 225 & 105 & 2.76 & 3.46 & 20.22 & 1 & 0 & 3 & 1 \\
#> 14.3 & 8 & 360 & 245 & 3.21 & 3.57 & 15.84 & 0 & 0 & 3 & 4 \\
#> 24.4 & 4 & 146.7 & 62 & 3.69 & 3.19 & 20 & 1 & 0 & 4 & 2 \\
#> 22.8 & 4 & 140.8 & 95 & 3.92 & 3.15 & 22.9 & 1 & 0 & 4 & 2 \\
#> 19.2 & 6 & 167.6 & 123 & 3.92 & 3.44 & 18.3 & 1 & 0 & 4 & 4 \\
#> 17.8 & 6 & 167.6 & 123 & 3.92 & 3.44 & 18.9 & 1 & 0 & 4 & 4 \\
#> 16.4 & 8 & 275.8 & 180 & 3.07 & 4.07 & 17.4 & 0 & 0 & 3 & 3 \\
#> 17.3 & 8 & 275.8 & 180 & 3.07 & 3.73 & 17.6 & 0 & 0 & 3 & 3 \\
#> 15.2 & 8 & 275.8 & 180 & 3.07 & 3.78 & 18 & 0 & 0 & 3 & 3 \\
#> 10.4 & 8 & 472 & 205 & 2.93 & 5.25 & 17.98 & 0 & 0 & 3 & 4 \\
#> 10.4 & 8 & 460 & 215 & 3 & 5.424 & 17.82 & 0 & 0 & 3 & 4 \\
#> 14.7 & 8 & 440 & 230 & 3.23 & 5.345 & 17.42 & 0 & 0 & 3 & 4 \\
#> 32.4 & 4 & 78.7 & 66 & 4.08 & 2.2 & 19.47 & 1 & 1 & 4 & 1 \\
#> 30.4 & 4 & 75.7 & 52 & 4.93 & 1.615 & 18.52 & 1 & 1 & 4 & 2 \\
#> 33.9 & 4 & 71.1 & 65 & 4.22 & 1.835 & 19.9 & 1 & 1 & 4 & 1 \\
#> 21.5 & 4 & 120.1 & 97 & 3.7 & 2.465 & 20.01 & 1 & 0 & 3 & 1 \\
#> 15.5 & 8 & 318 & 150 & 2.76 & 3.52 & 16.87 & 0 & 0 & 3 & 2 \\
#> 15.2 & 8 & 304 & 150 & 3.15 & 3.435 & 17.3 & 0 & 0 & 3 & 2 \\
#> 13.3 & 8 & 350 & 245 & 3.73 & 3.84 & 15.41 & 0 & 0 & 3 & 4 \\
#> 19.2 & 8 & 400 & 175 & 3.08 & 3.845 & 17.05 & 0 & 0 & 3 & 2 \\
#> 27.3 & 4 & 79 & 66 & 4.08 & 1.935 & 18.9 & 1 & 1 & 4 & 1 \\
#> 26 & 4 & 120.3 & 91 & 4.43 & 2.14 & 16.7 & 0 & 1 & 5 & 2 \\
#> 30.4 & 4 & 95.1 & 113 & 3.77 & 1.513 & 16.9 & 1 & 1 & 5 & 2 \\
#> 15.8 & 8 & 351 & 264 & 4.22 & 3.17 & 14.5 & 0 & 1 & 5 & 4 \\
#> 19.7 & 6 & 145 & 175 & 3.62 & 2.77 & 15.5 & 0 & 1 & 5 & 6 \\
#> 15 & 8 & 301 & 335 & 3.54 & 3.57 & 14.6 & 0 & 1 & 5 & 8 \\
#> 21.4 & 4 & 121 & 109 & 4.11 & 2.78 & 18.6 & 1 & 1 & 4 & 2 \\
#> \end{longtable}
#> \end{center}